+++ HISTORY.rst
@@ -5,6 +5,7 @@ History
 1.2.7 (unreleased)
 ------------------
 
+- Create model when creating new schematisation (nens/rana#3533)
 - Add cancel button for running simulations (nens/rana-qgis-plugin#235)